+++ src/bittytax/conv/parsers/__init__.py
@@ -58,6 +58,7 @@ from . import (
     hitbtc,
     hotbit,
     ii,
+    jupiter,
     kinesis,
     koinly,
     kraken,