+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Conversion tool: added parser for BloFin exchange.
 - Coinbase parser: added "Subscription", "Prime Deposit", "Transfer", "Retail Staking Transfer" and "Retail Unstaking Transfer" transaction types.
 - Conversion tool: added parser for Coinbase Prime exchange.
+- Conversion tool: added parser for Jupiter exchange.
 ### Changed
 - Config: fiat_income to True.
 - Price tool: CoinDesk API deprecated.